django-api/pyproject.toml
[project]
name = "django-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
